@@ -1,5 +1,5 @@
 import { isDevMode, NgModule } from '@angular/core';
-import { ServerModule, ServerTransferStateModule } from '@angular/platform-server';
+import { ServerModule, } from '@angular/platform-server';
 
 import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
@@ -9,7 +9,6 @@ import { APP_BASE_HREF } from '@angular/common';
   imports: [
     AppModule,
     ServerModule,
-    ServerTransferStateModule
   ],
   providers: [
     { provide: APP_BASE_HREF, useFactory: () => isDevMode() ? '/us-central1/ssr' : '/ssr' },
