@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { ServerModule, ServerTransferStateModule } from '@angular/platform-server';
+import { ServerModule } from '@angular/platform-server';
 
 import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
@@ -8,7 +8,6 @@ import { AppComponent } from './app.component';
   imports: [
     AppModule,
     ServerModule,
-    ServerTransferStateModule,
   ],
   bootstrap: [AppComponent],
 })
